--- Tests/Testably.Abstractions.Tests/FileSystem/Directory/Exists/DirectoryExistsTests.cs
@@ -1,26 +1,0 @@
-namespace Testably.Abstractions.Tests.FileSystem.Directory.Exists;
-
-// ReSharper disable once PartialTypeWithSinglePart
-public abstract partial class DirectoryExistsTests<TFileSystem>
-	: FileSystemTestBase<TFileSystem>
-	where TFileSystem : IFileSystem
-{
-	[SkippableTheory]
-	[AutoData]
-	public void Exists_File_ShouldReturnFalse(string path)
-	{
-		FileSystem.File.WriteAllText(path, null);
-
-		bool result = FileSystem.Directory.Exists(path);
-
-		result.Should().BeFalse();
-	}
-
-	[SkippableFact]
-	public void Exists_Null_ShouldReturnFalse()
-	{
-		bool result = FileSystem.Directory.Exists(null);
-
-		result.Should().BeFalse();
-	}
-}
