+++ Tests/Testably.Abstractions.Tests/FileSystem/DirectoryInfo/DeleteTests.cs
@@ -1,0 +1,96 @@
+using System.IO;
+using Testably.Abstractions.FileSystem;
+
+namespace Testably.Abstractions.Tests.FileSystem.DirectoryInfo;
+
+// ReSharper disable once PartialTypeWithSinglePart
+public abstract partial class DeleteTests<TFileSystem>
+	: FileSystemTestBase<TFileSystem>
+	where TFileSystem : IFileSystem
+{
+	[SkippableTheory]
+	[AutoData]
+	public void Delete_MissingDirectory_ShouldDeleteDirectory(string path)
+	{
+		IDirectoryInfo sut = FileSystem.DirectoryInfo.New(path);
+		sut.Exists.Should().BeFalse();
+
+		Exception? exception = Record.Exception(() =>
+		{
+			sut.Delete();
+		});
+
+		exception.Should().BeOfType<DirectoryNotFoundException>()
+		   .Which.Message.Should()
+		   .Be($"Could not find a part of the path '{sut.FullName}'.");
+	}
+
+	[SkippableTheory]
+	[AutoData]
+	public void Delete_Recursive_WithSubdirectory_ShouldDeleteDirectoryWithContent(
+		string path, string subdirectory)
+	{
+		string subdirectoryPath = FileSystem.Path.Combine(path, subdirectory);
+		FileSystem.Directory.CreateDirectory(subdirectoryPath);
+		IDirectoryInfo sut = FileSystem.DirectoryInfo.New(path);
+		sut.Exists.Should().BeTrue();
+
+		sut.Delete(true);
+
+#if NETFRAMEWORK
+		// The DirectoryInfo is not updated in .NET Framework!
+		sut.Exists.Should().BeTrue();
+#else
+		sut.Exists.Should().BeFalse();
+#endif
+		FileSystem.Directory.Exists(sut.FullName).Should().BeFalse();
+		FileSystem.Directory.Exists(subdirectoryPath).Should().BeFalse();
+	}
+
+	[SkippableTheory]
+	[AutoData]
+	public void Delete_ShouldDeleteDirectory(string path)
+	{
+		FileSystem.Directory.CreateDirectory(path);
+		IDirectoryInfo sut = FileSystem.DirectoryInfo.New(path);
+		sut.Exists.Should().BeTrue();
+
+		sut.Delete();
+
+#if NETFRAMEWORK
+		// The DirectoryInfo is not updated in .NET Framework!
+		sut.Exists.Should().BeTrue();
+#else
+		sut.Exists.Should().BeFalse();
+#endif
+		FileSystem.Directory.Exists(sut.FullName).Should().BeFalse();
+	}
+
+	[SkippableTheory]
+	[AutoData]
+	public void Delete_WithSubdirectory_ShouldNotDeleteDirectory(
+		string path, string subdirectory)
+	{
+		FileSystem.Directory.CreateDirectory(FileSystem.Path.Combine(path, subdirectory));
+		IDirectoryInfo sut = FileSystem.DirectoryInfo.New(path);
+		sut.Exists.Should().BeTrue();
+
+		Exception? exception = Record.Exception(() =>
+		{
+			sut.Delete();
+		});
+
+		exception.Should().BeOfType<IOException>();
+#if !NETFRAMEWORK
+		if (Test.RunsOnWindows)
+		{
+			// Path information only included in exception message on Windows and not in .NET Framework
+			exception.Should().BeOfType<IOException>()
+			   .Which.Message.Should().Contain($"'{sut.FullName}'");
+		}
+#endif
+
+		sut.Exists.Should().BeTrue();
+		FileSystem.Directory.Exists(sut.FullName).Should().BeTrue();
+	}
+}
