--- Tests/Testably.Abstractions.Tests/FileSystem/DirectoryInfo/GetDirectories/DirectoryInfoGetDirectoriesTests.cs
@@ -1,171 +1,0 @@
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using Testably.Abstractions.FileSystem;
-using Testably.Abstractions.Testing.FileSystemInitializer;
-
-namespace Testably.Abstractions.Tests.FileSystem.DirectoryInfo.GetDirectories;
-
-// ReSharper disable once PartialTypeWithSinglePart
-public abstract partial class DirectoryInfoGetDirectoriesTests<TFileSystem>
-	: FileSystemTestBase<TFileSystem>
-	where TFileSystem : IFileSystem
-{
-	[SkippableTheory]
-	[AutoData]
-	public void
-		GetDirectories_SearchOptionAllDirectories_ShouldReturnAllSubdirectories(
-			string path)
-	{
-		IFileSystemDirectoryInitializer<TFileSystem> initialized =
-			FileSystem.Initialize()
-			   .WithSubdirectory(path).Initialized(s => s
-				   .WithSubdirectory("foo/xyz")
-				   .WithSubdirectory("bar"));
-		IDirectoryInfo baseDirectory =
-			(IDirectoryInfo)initialized[0];
-
-		IDirectoryInfo[] result = baseDirectory
-		   .GetDirectories("*", SearchOption.AllDirectories);
-
-		result.Length.Should().Be(3);
-		result.Should().Contain(d => d.Name == "foo");
-		result.Should().Contain(d => d.Name == "bar");
-		result.Should().Contain(d => d.Name == "xyz");
-	}
-
-	[SkippableTheory]
-#if NETFRAMEWORK
-	[InlineAutoData(false, "")]
-#else
-	[InlineAutoData(true, "")]
-#endif
-	[InlineAutoData(true, "*")]
-	[InlineAutoData(true, ".")]
-	[InlineAutoData(true, "*.*")]
-	[InlineData(true, "a*c", "abc")]
-	[InlineData(true, "ab*c", "abc")]
-	[InlineData(true, "abc?", "abc")]
-	[InlineData(false, "ab?c", "abc")]
-	[InlineData(false, "ac", "abc")]
-	public void GetDirectories_SearchPattern_ShouldReturnExpectedValue(
-		bool expectToBeFound, string searchPattern, string subdirectoryName)
-	{
-		IDirectoryInfo baseDirectory =
-			FileSystem.Directory.CreateDirectory("foo");
-		baseDirectory.CreateSubdirectory(subdirectoryName);
-
-		IDirectoryInfo[] result = baseDirectory
-		   .GetDirectories(searchPattern);
-
-		if (expectToBeFound)
-		{
-			result.Should().ContainSingle(d => d.Name == subdirectoryName,
-				$"it should match '{searchPattern}'");
-		}
-		else
-		{
-			result.Should()
-			   .BeEmpty($"{subdirectoryName} should not match '{searchPattern}'");
-		}
-	}
-
-#if FEATURE_FILESYSTEM_ENUMERATION_OPTIONS
-	[SkippableTheory]
-	[AutoData]
-	public void
-		GetDirectories_WithEnumerationOptions_ShouldConsiderSetOptions(
-			string path)
-	{
-		IDirectoryInfo baseDirectory =
-			FileSystem.Directory.CreateDirectory(path);
-		baseDirectory.CreateSubdirectory("foo/xyz");
-		baseDirectory.CreateSubdirectory("bar");
-
-		IDirectoryInfo[] result = baseDirectory
-		   .GetDirectories("XYZ",
-				new EnumerationOptions
-				{
-					MatchCasing = MatchCasing.CaseInsensitive,
-					RecurseSubdirectories = true,
-					// Filename could start with a leading '.' indicating it as Hidden in Linux
-					AttributesToSkip = FileAttributes.System
-				});
-
-		result.Length.Should().Be(1);
-		result.Should().NotContain(d => d.Name == "foo");
-		result.Should().Contain(d => d.Name == "xyz");
-		result.Should().NotContain(d => d.Name == "bar");
-	}
-#endif
-
-	[SkippableTheory]
-	[AutoData]
-	public void GetDirectories_WithNewline_ShouldThrowArgumentException(
-		string path)
-	{
-		IDirectoryInfo baseDirectory =
-			FileSystem.DirectoryInfo.New(path);
-		string searchPattern = "foo\0bar";
-
-		Exception? exception = Record.Exception(() =>
-		{
-			_ = baseDirectory.GetDirectories(searchPattern).FirstOrDefault();
-		});
-
-		exception.Should().BeOfType<ArgumentException>();
-	}
-
-	[SkippableTheory]
-	[AutoData]
-	public void
-		GetDirectories_WithoutSearchString_ShouldReturnAllDirectSubdirectories(
-			string path)
-	{
-		IDirectoryInfo baseDirectory =
-			FileSystem.Directory.CreateDirectory(path);
-		baseDirectory.CreateSubdirectory("foo/xyz");
-		baseDirectory.CreateSubdirectory("bar");
-
-		IDirectoryInfo[] result = baseDirectory
-		   .GetDirectories();
-
-		result.Length.Should().Be(2);
-		result.Should().Contain(d => d.Name == "foo");
-		result.Should().NotContain(d => d.Name == "xyz");
-		result.Should().Contain(d => d.Name == "bar");
-	}
-
-	[SkippableTheory]
-	[AutoData]
-	public void GetDirectories_WithSearchPattern_ShouldReturnMatchingSubdirectory(
-		string path)
-	{
-		IDirectoryInfo baseDirectory =
-			FileSystem.Directory.CreateDirectory(path);
-		baseDirectory.CreateSubdirectory("foo");
-		baseDirectory.CreateSubdirectory("bar");
-
-		IEnumerable<IDirectoryInfo> result = baseDirectory
-		   .GetDirectories("foo");
-
-		result.Should().ContainSingle(d => d.Name == "foo");
-	}
-
-	[SkippableTheory]
-	[AutoData]
-	public void
-		GetDirectories_WithSearchPatternInSubdirectory_ShouldReturnMatchingSubdirectory(
-			string path)
-	{
-		IDirectoryInfo baseDirectory =
-			FileSystem.Directory.CreateDirectory(path);
-		baseDirectory.CreateSubdirectory("foo/xyz");
-		baseDirectory.CreateSubdirectory("bar/xyz");
-
-		IEnumerable<IDirectoryInfo> result = baseDirectory
-		   .GetDirectories("xyz", SearchOption.AllDirectories);
-
-		result.Count().Should().Be(2);
-	}
-}
